20240829_오경석_analysis테이블 analResultImg_Name 컬럼 LONGTEXT 타입으로 변경

+++ src/main/java/com/smhrd/repository/AnalysisRepository.java
@@ -9,6 +9,6 @@ import org.springframework.data.repository.query.Param;
 import com.smhrd.entity.Analysis;
 
 
-public interface AnalysisRepository extends JpaRepository<Analysis, Long> {
+public interface AnalysisRepository extends JpaRepository<Analysis, Integer> {
 
 }
